chore: refactor test folder
